More applying of rules. 404 and 405 now handled

+++ src/utils/State.hpp
@@ -16,6 +16,7 @@
 #include <iostream>
 #include <vector>
 #include <map>
+#include <cstddef>
 
 struct  CGI;
 struct  Location;
@@ -73,9 +74,9 @@ struct  Location: public IBlock
 {  
     std::string _root; // where it's getting redirected;
     std::string _path; // the redirect 
-    std::vector<std::string>    _index;
-    std::vector<std::string>    _methods;
-    unsigned long long   _max_body_size;
+    std::vector<std::string>	_index;
+    std::vector<std::string>	_methods;
+    size_t					_max_body_size;
     bool    _autoindex;
     std::vector<CGI>    _cgi;
     Location();
